requirements.txt
aiohttp==3.8.4
PyYAML==6.0
Jinja2==3.1.2
aiohttp-jinja2==1.5.1
redis==4.5.1